requirements.txt
Flask==2.0.1
tensorflow==2.8.0
Pillow==9.0.1
numpy==1.22.3
scipy==1.8.0
matplotlib==3.5.1
tensorflow-hub==0.12.0
Werkzeug==2.0.2
